+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - CSV download of points plotted on map
 - CSV download of flag types added to metadata download
 - Ability to add annotation in flag option modal
+- Checkbox to use map option min/max years in graph options
 
 ### Changed